Clean up https
